requirements.txt
django==3.2.9
djangorestframework==3.12.4
pandas==1.1.5
numpy==1.19.5
sklearn==0.0